Egendefinerte farger på faser i fasevelgeren (#1615)
* Added dynamic darken and saturation for custom coloring of phases

* Restructuring

* Changelog

* Dynamically adjust font-color for higher luminance values
