+++ SharePointFramework/ProjectWebParts/src/components/ProjectPhases/ProjectPhase/ProjectPhase.module.scss
@@ -1,14 +1,16 @@
-$project-color: #89ce87;
-$project-current-color: #54b054;
-$project-current-color-hover: #397e39;
+$project-color: var(--phase-color, #89ce87);
+$project-current-color: var(--phase-current-color, #54b054);
+$project-current-color-hover: var(--phase-current-hover-color, #397e39);
 
-$portfolio-color: #9dceee;
-$portfolio-current-color: #64addb;
-$portfolio-current-color-hover: #5688af;
+$portfolio-color: var(--phase-color, #9dceee);
+$portfolio-current-color: var(--phase-current-color, #64addb);
+$portfolio-current-color-hover: var(--phase-current-hover-color, #5688af);
 
-$end-color: #f0c993;
-$end-current-color: #f09440;
-$end-current-color-hover: #ba6f29;
+$end-color: var(--phase-color, #f0c993);
+$end-current-color: var(--phase-current-color, #f09440);
+$end-current-color-hover: var(--phase-current-hover-color, #ba6f29);
+
+$font-color: var(--font-color, white);
 
 $arrow-font-size: 4rem;
 $arrow-height: 1em;
@@ -68,6 +70,7 @@ $project-phase-padding: 0.25em;
       border-top-color: inherit;
       border-bottom-color: inherit;
     }
+
     &::before {
       border-left-color: inherit;
     }
@@ -86,6 +89,7 @@ $project-phase-padding: 0.25em;
       border-top-color: inherit;
       border-bottom-color: inherit;
     }
+
     &::before {
       border-left-color: inherit;
     }
@@ -99,6 +103,7 @@ $project-phase-padding: 0.25em;
         border-top-color: inherit;
         border-bottom-color: inherit;
       }
+
       &::before {
         border-left-color: inherit;
       }
@@ -113,6 +118,7 @@ $project-phase-padding: 0.25em;
       border-top-color: inherit;
       border-bottom-color: inherit;
     }
+
     &::before {
       border-left-color: inherit;
     }
@@ -126,6 +132,7 @@ $project-phase-padding: 0.25em;
         border-top-color: inherit;
         border-bottom-color: inherit;
       }
+
       &::before {
         border-left-color: inherit;
       }
@@ -163,6 +170,7 @@ $project-phase-padding: 0.25em;
     border-top-color: inherit;
     border-bottom-color: inherit;
   }
+
   &::before {
     border-left-color: inherit;
   }
@@ -176,6 +184,7 @@ $project-phase-padding: 0.25em;
       border-top-color: inherit;
       border-bottom-color: inherit;
     }
+
     &::before {
       border-left-color: inherit;
     }
@@ -189,6 +198,7 @@ $project-phase-padding: 0.25em;
       border-top-color: inherit;
       border-bottom-color: inherit;
     }
+
     &::before {
       border-left-color: inherit;
     }
@@ -201,6 +211,7 @@ $project-phase-padding: 0.25em;
         border-top-color: inherit;
         border-bottom-color: inherit;
       }
+
       &::before {
         border-left-color: inherit;
       }
@@ -215,6 +226,7 @@ $project-phase-padding: 0.25em;
       border-top-color: inherit;
       border-bottom-color: inherit;
     }
+
     &::before {
       border-left-color: inherit;
     }
@@ -227,6 +239,7 @@ $project-phase-padding: 0.25em;
         border-top-color: inherit;
         border-bottom-color: inherit;
       }
+
       &::before {
         border-left-color: inherit;
       }
@@ -238,7 +251,7 @@ $project-phase-padding: 0.25em;
   font-weight: 400;
   padding: $chevron-width;
   text-align: center;
-  color: white;
+  color: $font-color;
 
   span {
     cursor: pointer;
